requirements.txt
GitPython==1.0.1
gitdb==0.6.4
smmap==0.9.0
stomp.py==4.1.2
wsgiref==0.1.2

#sudo apt-get install -y python-pip


#sudo pip install GitPython==1.0.1
 #sudo pip install gitdb==0.6.4
 #sudo pip install smmap==0.9.0
  #sudo pip install stomp.py==4.1.2
  #sudo pip install wsgiref==0.1.2